Refactor wasmer memory into separate module
