+++ src/call_context.rs
@@ -21,6 +21,7 @@ pub trait Resolver: Invoke + ModuleImportResolver + Clone + Default {}
 
 pub use crate::resolver::CompoundResolver as StandardABI;
 use crate::resolver::{HostImportsResolver, Env, ImportReference};
+use crate::memory::WasmerMemory;
 
 use wasmer::{imports, wat2wasm, Function, Instance, Module, NativeFunc, Store, ImportObject, Exports, Memory, LazyInit};
 use wasmer_compiler_cranelift::Cranelift;
@@ -94,84 +95,6 @@ pub trait Invoke: Sized {
     ) -> Result<Option<RuntimeValue>, VMError>;
 }
 
-/// Check that the given offset and length fits into the memory bounds. If not,
-/// it will try to grow the memory.
-fn check_bounds(memory: &Memory, offset: u64, len: usize) -> Result<(), VMError> {
-    if memory.data_size() < offset + len as u64 {
-        let cur_pages = memory.size().0;
-        let capacity = cur_pages as usize * wasmer::WASM_PAGE_SIZE;
-        let missing = offset as usize + len - capacity;
-        // Ceiling division
-        let req_pages = ((missing + wasmer::WASM_PAGE_SIZE - 1)
-            / wasmer::WASM_PAGE_SIZE) as u32;
-        memory.grow(req_pages).map_err(VMError::MemoryNotFound)?; // todo this grow will probably need to go away
-    }
-    Ok(())
-}
-
-/// Read bytes from memory at the given offset and length
-fn read_memory_bytes(
-    memory: &Memory,
-    offset: u64,
-    len: usize,
-) -> Result<Vec<u8>, VMError> {
-    check_bounds(memory, offset, len)?;
-    let offset = offset as usize;
-    let vec: Vec<_> = memory.view()[offset..(offset + len)]
-        .iter()
-        .map(|cell| cell.get())
-        .collect();
-    Ok(vec)
-}
-
-/// Write bytes into memory at the given offset
-fn write_memory_bytes(
-    memory: &Memory,
-    offset: u64,
-    bytes: impl AsRef<[u8]>,
-) -> Result<(), VMError> {
-    let slice = bytes.as_ref();
-    let len = slice.len();
-    check_bounds(memory, offset, len as _)?;
-    let offset = offset as usize;
-    memory.view()[offset..(offset + len)]
-        .iter()
-        .zip(slice.iter())
-        .for_each(|(cell, v)| cell.set(*v));
-    Ok(())
-}
-
-pub struct WasmerMemory {
-    inner: LazyInit<Memory>,
-}
-
-impl WasmerMemory {
-    pub fn init_env_memory(
-        &mut self,
-        exports: &wasmer::Exports,
-    ) -> std::result::Result<(), VMError> {
-        let memory = exports.get_memory("memory")?;
-        self.inner.initialize(memory.clone());
-        Ok(())
-    }
-}
-
-
-// impl WasmerMemory {
-//     pub fn new(memory_ref: &wasmer::Memory) -> WasmerMemoryRef {
-//         WasmerMemoryRef { memory_ref }
-//     }
-//     pub fn with_direct_access<R, F: FnOnce(&[u8]) -> R>(&self, f: F) -> R {
-//         let buf = unsafe { self.memory_ref.data_unchecked() };
-//         f(buf)
-//     }
-//
-//     pub fn with_direct_access_mut<R, F: FnOnce(&mut [u8]) -> R>(&mut self, f: F) -> R {
-//         let buf = unsafe { self.memory_ref.data_unchecked_mut() };
-//         f(buf)
-//     }
-//
-// }
 
 pub struct CallContext<'a> {
     state: &'a mut NetworkState,
@@ -258,8 +181,8 @@ impl<'a> CallContext<'a> {
 
             let mut wasmer_memory = WasmerMemory { inner: LazyInit::new() };
             wasmer_memory.init_env_memory(&wasmer_instance.exports)?;
-            unsafe { write_memory_bytes(wasmer_memory.inner.get_unchecked(), 0, contract.state().as_bytes()) };
-            unsafe { write_memory_bytes(wasmer_memory.inner.get_unchecked(), contract.state().as_bytes().len() as u64, query.as_bytes()) };
+            unsafe { WasmerMemory::write_memory_bytes(wasmer_memory.inner.get_unchecked(), 0, contract.state().as_bytes()) };
+            unsafe { WasmerMemory::write_memory_bytes(wasmer_memory.inner.get_unchecked(), contract.state().as_bytes().len() as u64, query.as_bytes()) };
 
             self.stack
                 .push(StackFrame::new_query(target, wasmer_memory, query));
@@ -288,7 +211,7 @@ impl<'a> CallContext<'a> {
         // WASMER
         let mut wasmer_memory = WasmerMemory { inner: LazyInit::new() };
         wasmer_memory.init_env_memory(&wasmer_instance.exports)?;
-        let read_buffer = unsafe { read_memory_bytes(wasmer_memory.inner.get_unchecked(), 0, wasmer_memory.inner.get_unchecked().data_size() as usize)? };
+        let read_buffer = unsafe { WasmerMemory::read_memory_bytes(wasmer_memory.inner.get_unchecked(), 0, wasmer_memory.inner.get_unchecked().data_size() as usize)? };
         let mut source = Source::new(&read_buffer);
         let result = ReturnValue::decode(&mut source).expect("query result decoded");
         self.stack.pop();
@@ -354,8 +277,8 @@ impl<'a> CallContext<'a> {
             // WASMER
             let mut wasmer_memory = WasmerMemory { inner: LazyInit::new() };
             wasmer_memory.init_env_memory(&wasmer_instance.exports)?;
-            unsafe { write_memory_bytes(wasmer_memory.inner.get_unchecked(), 0, contract.state().as_bytes()) };
-            unsafe { write_memory_bytes(wasmer_memory.inner.get_unchecked(), contract.state().as_bytes().len() as u64, transaction.as_bytes()) };
+            unsafe { WasmerMemory::write_memory_bytes(wasmer_memory.inner.get_unchecked(), 0, contract.state().as_bytes()) };
+            unsafe { WasmerMemory::write_memory_bytes(wasmer_memory.inner.get_unchecked(), contract.state().as_bytes().len() as u64, transaction.as_bytes()) };
 
             self.stack.push(StackFrame::new_transaction(
                 target,
@@ -397,7 +320,7 @@ impl<'a> CallContext<'a> {
             // WASMER
             let mut wasmer_memory = WasmerMemory { inner: LazyInit::new() };
             wasmer_memory.init_env_memory(&wasmer_instance.exports)?;
-            let read_buffer = unsafe { read_memory_bytes(wasmer_memory.inner.get_unchecked(), 0, wasmer_memory.inner.get_unchecked().data_size() as usize)? };
+            let read_buffer = unsafe { WasmerMemory::read_memory_bytes(wasmer_memory.inner.get_unchecked(), 0, wasmer_memory.inner.get_unchecked().data_size() as usize)? };
             let mut source = Source::new(&read_buffer);
             let state = ContractState::decode(&mut source).expect("query result decoded");
             *(*contract).state_mut() = state;
